Add support for MySQL
This commit adds support for MySQL. The main change is using the adapter
name in Base to determine whether we're using PostgreSQL instead of
referencing the class. The reason is the class might be absent in MySQL
setups and result in a runtime error trying to reference an undefined
constant.

The other changes are to two test files - we need to skip some tests
that don't make sense for MySQL.