Fix index comparison in extraneous_indexes
This commit fixes a bug that caused incompatible indexes to be assumed
to cover one another. For example, a GIST and a hash index on the same
column would be assumed to be EXACTLY the same index because index
type was not taken into account.

From now on, the following four properties are taken into account: type,
method, operator class, and WHERE clause.