Fix test errors due to non-derministic output
This commit fixes a problem with the test suite that caused it to fail
from time to time due to non-deterministic order of error messages
produced by detectors. The order depends on the order in which models
and/or tables are traversed which is non-deterministic itself.

The problem is solved by always sorting expected and actual warnings
produced by detectors before comparing them for equality.