Fix ignore_columns in missing_unique_indexes
This commit fixes a problem that made it impossible to ignore the right column
when working with a has_one association. The root cause was looking at the model
containing the has_one relationship instead of the one being referred to, which
is the model whose table contains the actual foreign key.